Add Constants of popular routes. City.record for database management. Default 'false' fields on RequestTicketsDTO. UkrPlacesRepository -> id to Long

+++ src/main/java/com/booking/app/repositories/UkrPlacesRepository.java
@@ -4,14 +4,14 @@ import com.booking.app.entity.UkrainianPlaces;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UkrPlacesRepository extends JpaRepository<UkrainianPlaces, BigInteger> {
+public interface UkrPlacesRepository extends JpaRepository<UkrainianPlaces, Long> {
 
-    Optional<List<UkrainianPlaces>> findUkrainianPlacesByNameEngStartsWithIgnoreCaseAndNameEngNotContainingIgnoreCase(String startLetters,String exclusion);
-    Optional<List<UkrainianPlaces>> findUkrainianPlacesByNameUaStartsWithIgnoreCaseAndNameUaNotContainingIgnoreCase(String startLetters,String exclusion);
+    Optional<List<UkrainianPlaces>> findUkrainianPlacesByNameEngStartsWithIgnoreCaseAndNameEngNotContainingIgnoreCase(String startLetters, String exclusion);
+
+    Optional<List<UkrainianPlaces>> findUkrainianPlacesByNameUaStartsWithIgnoreCaseAndNameUaNotContainingIgnoreCase(String startLetters, String exclusion);
 
 }
